# Optimize PointLight2D shadow rendering by reducing draw calls and RD state changes